Refactor drawing utilities from tree editor
