+++ src/components/D3BehaviorTreeEditor.vue
@@ -30,34 +30,21 @@
 <script setup lang="ts">
 import { EditorSkin, useEditorStore } from '@/stores/editor'
 import { useEditNodeStore } from '@/stores/edit_node'
-import { useROSStore } from '@/stores/ros'
-import type {
-  DataEdgePoint,
-  DataEdgeTerminal,
-  DataEdge,
-  DocumentedNode,
-  DropTarget,
-  NodeDataLocation,
-  OptionData,
-  BTEditorNode,
-  TrimmedNodeData,
-  NodeStructure,
-  NodeOption,
-  NodeIO,
-  Wiring,
-  UUIDString
+import {
+  type DataEdgePoint,
+  type DataEdgeTerminal,
+  type DataEdge,
+  type DropTarget,
+  type BTEditorNode,
+  type UUIDString,
+  NodeStateValues
 } from '@/types/types'
-import { Position, IOKind, NodeStateValues } from '@/types/types'
-import { getDefaultValue, prettyprint_type, serializeNodeOptions, typesCompatible } from '@/utils'
+import { IOKind } from '@/types/types'
 import { notify } from '@kyvg/vue3-notification'
 import * as d3 from 'd3'
-import { onMounted, ref, watch, watchEffect } from 'vue'
-import { addNode, moveNode, replaceNode } from '@/tree_manipulation'
-import type { HierarchyNode, HierarchyLink } from 'd3-hierarchy'
-import { flextree, type FlextreeNode } from 'd3-flextree'
-import * as uuid from 'uuid'
-import { rosToUuid, uuidToRos } from '@/utils'
-import type { WireNodeDataRequest, WireNodeDataResponse } from '@/types/services/WireNodeData'
+import { onMounted, ref, watch } from 'vue'
+import type { FlextreeNode } from 'd3-flextree'
+import { drawDataLine } from '@/tree_display/draw_tree_data'
 import {
   faBolt,
   faCheck,
@@ -66,21 +53,37 @@ import {
   faPowerOff,
   faXmark
 } from '@fortawesome/free-solid-svg-icons'
+import {
+  data_edge_css_class,
+  data_graph_comaptible_css_class,
+  data_graph_select_css_class,
+  data_edge_highlight_css_id,
+  data_vert1_highlight_css_id,
+  data_vert2_highlight_css_id,
+  data_vert_group_css_class,
+  data_vert_label_css_class,
+  drop_target_css_class,
+  io_gripper_size,
+  node_body_css_class,
+  node_selected_css_class,
+  tree_node_css_class,
+  node_state_css_class,
+  icon_width,
+  node_padding
+} from '@/tree_display/draw_tree_config'
+import { D3TreeDisplay } from '@/tree_display/draw_tree'
+import { findNodeInTreeList, getNodeStates } from '@/tree_selection'
 
 const editor_store = useEditorStore()
 const edit_node_store = useEditNodeStore()
-const ros_store = useROSStore()
 
 const viewport_ref = ref<SVGSVGElement>()
 const svg_g_ref = ref<SVGGElement>()
-const g_vertices_ref = ref<SVGGElement>()
-const g_edges_ref = ref<SVGGElement>()
-const g_data_graph_ref = ref<SVGGElement>()
-const g_data_edges_ref = ref<SVGGElement>()
-const g_data_vertices_ref = ref<SVGGElement>()
-const g_drop_targets_ref = ref<SVGGElement>()
+const tree_root_ref = ref<SVGGElement>()
 const draw_indicator_ref = ref<SVGPathElement>()
 const selection_rect_ref = ref<SVGRectElement>()
+
+let tree_display: D3TreeDisplay | undefined = undefined
 
 let zoomObject: d3.ZoomBehavior<SVGSVGElement, unknown> | undefined = undefined
 
@@ -95,55 +98,6 @@ let pan_direction: number[] = [0.0, 0.0]
 const pan_rate: number = 30
 const drag_pan_boundary: number = 50
 const pan_per_frame: number = 10.0
-
-const line_wrap_regex: RegExp = /[a-z0-9][A-Z]|[_\- ][a-zA-Z]/dg
-
-const io_gripper_size: number = 15
-const io_gripper_spacing: number = 10
-
-const io_edge_offset: number = 20
-const io_edge_bump_thresh: number = 50
-const io_edge_bump_factor: number = 0.02
-const io_edge_curve_offset: number = 5
-const io_edge_curve_factor: number = 0.0001
-
-const node_padding: number = 10
-const node_spacing: number = 80
-const name_line_length: number = 20
-const name_first_line_indent: number = 3
-const class_line_length: number = 25
-const icon_width: number = 30
-const node_name_height: number = 40
-const node_class_height: number = 30
-const drop_target_root_size: number = 150
-
-// constants for css classes & ids used with d3
-const tree_node_css_class: string = 'node'
-const node_body_css_class: string = 'btnode'
-const node_name_css_class: string = 'node_name'
-const node_class_css_class: string = 'class_name'
-const node_state_css_class: string = 'state_icon'
-const tree_edge_css_class: string = 'link'
-const drop_target_css_class: string = 'drop_target'
-const data_vert_group_css_class: string = 'gripper-group'
-const data_vert_grip_css_class: string = 'gripper'
-const data_vert_label_css_class: string = 'label'
-const data_vert_label_type_css_class: string = 'label-type'
-const data_edge_css_class: string = 'data-link'
-
-const data_edge_highlight_css_id: string = 'hightlightEdge'
-const data_vert1_highlight_css_id: string = 'highlightV1'
-const data_vert2_highlight_css_id: string = 'highlightV2'
-
-const data_graph_hover_css_class: string = 'data-hover'
-const data_graph_comaptible_css_class: string = 'compatible'
-
-const node_selected_css_class: string = 'node-selected'
-const data_graph_select_css_class: string = 'data-select'
-
-// This is the base transition for tree updates
-// Use <selection>.transition(tree_transition)
-let tree_transition: d3.Transition<d3.BaseType, unknown, null, undefined>
 
 function resetView() {
   if (
@@ -163,28 +117,6 @@ function resetView() {
   const offset = (height * 0.5) / zoom_factor - 60.0
 
   viewport.call(zoomObject.translateTo, 0.0, offset)
-}
-
-function buildNodeMessage(node: DocumentedNode): NodeStructure {
-  const options = node.options.map((opt: NodeOption) => {
-    return {
-      key: opt.key,
-      value: getDefaultValue(prettyprint_type(opt.serialized_type), node.options)
-    } as OptionData
-  })
-
-  return {
-    node_id: uuidToRos(uuid.v4()),
-    name: node.node_class,
-    module: node.module,
-    node_class: node.node_class,
-    version: '',
-    max_children: 0,
-    child_ids: [],
-    options: serializeNodeOptions(options),
-    inputs: [],
-    outputs: []
-  }
 }
 
 function canvasMouseMovePanHandler(event: Event) {
@@ -259,179 +191,37 @@ function dragPanTimerHandler() {
   d3.select(viewport_ref.value!).call(zoomObject!.translateBy, pan_direction[0], pan_direction[1])
 }
 
-watch(
-  [() => editor_store.current_tree.structure, () => editor_store.is_layer_mode],
-  drawEverything,
-  { immediate: true }
-)
+watch([() => editor_store.tree_structure_list, () => editor_store.is_layer_mode], drawEverything, {
+  immediate: true
+})
 function drawEverything() {
-  if (svg_g_ref.value === undefined || g_vertices_ref.value === undefined) {
+  if (
+    svg_g_ref.value === undefined ||
+    tree_root_ref.value === undefined ||
+    draw_indicator_ref.value === undefined
+  ) {
     console.warn('DOM is broken')
     return
   }
 
-  if (editor_store.current_tree.structure === undefined) {
-    console.warn('Nothing to draw')
-    return
-  }
+  const root_element = d3.select(tree_root_ref.value)
 
   // Prepare transition config for synchronization, the typed select statement is necessary to give the transition appropriate typing
-  tree_transition = d3
+  const tree_transition = d3
     .select<d3.BaseType, unknown>(svg_g_ref.value)
     .transition('tree-transition')
     .duration(100)
     .ease(d3.easeQuad)
 
-  // Strips potentially additional properties
-  const onlyKeyAndType = (nodeData: NodeIO) =>
-    ({
-      key: nodeData.key,
-      serialized_type: nodeData.serialized_type
-    } as TrimmedNodeData)
+  tree_display = new D3TreeDisplay(
+    editor_store.selected_tree,
+    !editor_store.has_selected_subtree,
+    draw_indicator_ref.value,
+    root_element,
+    tree_transition
+  )
 
-  // Trim the serialized data values from the node data - we won't
-  // render them, so don't clutter the DOM with the data
-  const editor_nodes: BTEditorNode[] = editor_store.current_tree.structure.nodes.map((node) => {
-    return {
-      node_id: rosToUuid(node.node_id),
-      name: node.name,
-      node_class: node.node_class,
-      module: node.module,
-      max_children: node.max_children,
-      child_ids: node.child_ids.map(rosToUuid),
-      options: node.options.map(onlyKeyAndType),
-      inputs: node.inputs.map(onlyKeyAndType),
-      outputs: node.outputs.map(onlyKeyAndType),
-      size: { width: 1, height: 1 },
-      offset: { x: 0, y: 0 }
-    } as BTEditorNode
-  })
-
-  const forest_root: BTEditorNode = {
-    node_id: uuid.NIL,
-    name: '',
-    node_class: '',
-    module: '',
-    max_children: -1,
-    child_ids: [],
-    inputs: [],
-    outputs: [],
-    options: [],
-    size: { width: 0, height: 0 },
-    offset: { x: 0, y: 0 }
-  }
-
-  if (editor_nodes.findIndex((x) => x.node_id === uuid.NIL) < 0) {
-    editor_nodes.push(forest_root)
-  }
-
-  // Update the visual tree
-  const parents: Record<UUIDString, UUIDString> = {}
-  //const node_dict: Record<string, TrimmedNode> = {}; Is unused?
-  // Find parents for all nodes once
-  for (const i in editor_nodes) {
-    const node = editor_nodes[i]
-    //node_dict[node.name] = node;
-    for (const j in node.child_ids) {
-      parents[node.child_ids[j]] = node.node_id
-    }
-  }
-
-  const root: d3.HierarchyNode<BTEditorNode> = d3
-    .stratify<BTEditorNode>()
-    .id((node) => {
-      return node.node_id
-    })
-    .parentId((node) => {
-      // undefined if it has no parent - does that break the layout?
-      if (node.node_id in parents) {
-        return parents[node.node_id]
-      } else if (node.node_id === forest_root.node_id) {
-        return undefined
-      } else {
-        forest_root.child_ids.push(node.node_id)
-        return forest_root.node_id
-      }
-    })(editor_nodes)
-
-  root.sort(function (a, b) {
-    if (a.depth !== b.depth) {
-      return b.depth - a.depth
-    }
-    while (a.parent !== b.parent) {
-      a = a.parent!
-      b = b.parent!
-    }
-    const child_list = a.parent!.data.child_ids
-    return (
-      child_list.findIndex((x) => x === a.data.node_id) -
-      child_list.findIndex((x) => x === b.data.node_id)
-    )
-  })
-
-  const g_vertex = d3.select<SVGGElement, never>(g_vertices_ref.value)
-
-  const node = g_vertex
-    .selectAll<SVGSVGElement, d3.HierarchyNode<BTEditorNode>>('.' + tree_node_css_class)
-    .data(
-      root.descendants().filter((node) => node.data.node_id !== uuid.NIL),
-      (node) => node.id!
-    ) // Join performs enter, update and exit at once
-    .join(drawNewNodes)
-    .call(updateNodeBody)
-
-  updateNodeState()
-
-  // Since we want to return the tree, we can't use the .call() syntax here
-  const tree_layout = layoutTree(node, root)
-
-  drawEdges(tree_layout)
-  drawDropTargets(tree_layout)
-  drawDataGraph(tree_layout, editor_store.current_tree.structure.data_wirings)
-}
-
-function drawNewNodes(
-  selection: d3.Selection<d3.EnterElement, d3.HierarchyNode<BTEditorNode>, SVGGElement, never>
-) {
-  const group = selection
-    .append<SVGGElement>('g')
-    .classed(tree_node_css_class, true)
-    // The two css-classes below are currently unused
-    .classed('node--internal', (d) => d.children !== undefined && d.children.length > 0)
-    .classed('node--leaf', (d) => d.children === undefined || d.children.length == 0)
-    .on('click.select', (event, node: d3.HierarchyNode<BTEditorNode>) => {
-      if (event.shiftKey) {
-        edit_node_store.selectMultipleNodes([node.data.node_id])
-      } else {
-        edit_node_store.editorSelectionChange(node.data.node_id)
-      }
-      event.stopPropagation()
-    })
-
-  // No tree modifying if displaying a subtree
-  if (!editor_store.has_selected_subtree) {
-    group.on('mousedown.dragdrop', (event, node: d3.HierarchyNode<BTEditorNode>) => {
-      editor_store.startDraggingExistingNode(node)
-      event.stopPropagation()
-    })
-  }
-
-  group.append<SVGRectElement>('rect').classed(node_body_css_class, true)
-
-  group.append<SVGTextElement>('text').classed(node_name_css_class, true)
-
-  group.append<SVGTextElement>('text').classed(node_class_css_class, true)
-
-  group
-    .append('svg')
-    .classed(node_state_css_class, true)
-    .attr('width', icon_width)
-    .attr('height', icon_width)
-    .append('path')
-
-  // The join pattern requires a return of the appended elements
-  // For consistency the node body is filled using the update method
-  return group
+  tree_display.drawTree()
 }
 
 function getIcon(state: NodeStateValues | undefined) {
@@ -452,169 +242,25 @@ function getIcon(state: NodeStateValues | undefined) {
   }
 }
 
-function layoutText(element: SVGGElement, data: d3.HierarchyNode<BTEditorNode>): number {
-  // Track width of longest line and return that for box sizing
-  let max_width: number = 0
-
-  const group_elem = d3.select<SVGGElement, d3.HierarchyNode<BTEditorNode>>(element)
-
-  const name_elem = group_elem.select<SVGTextElement>('.' + node_name_css_class)
-  const class_elem = group_elem.select<SVGTextElement>('.' + node_class_css_class)
-
-  const node_name = data.data.name
-  const node_class = data.data.node_class
-
-  name_elem.selectAll<SVGTSpanElement, never>('tspan').remove()
-
-  class_elem.selectAll<SVGTSpanElement, never>('tspan').remove()
-
-  let title_lines: number = 0
-
-  // Find positions for potential line breaks
-  let wrap_indices: number[] = [0]
-  for (const match of node_name.matchAll(line_wrap_regex)) {
-    wrap_indices.push(match.index + 1)
-  }
-  wrap_indices.push(node_name.length)
-  wrap_indices.reverse()
-
-  // Place text into multiple lines
-  let current_index: number = 0
-  while (current_index < node_name.length) {
-    // Prepare DOM-element for next line
-    const tspan = name_elem.append('tspan')
-    tspan.attr('x', 0)
-    if (current_index > 0) {
-      tspan.attr('dy', node_name_height)
-    }
-
-    // Since this predicate is guaranteed to hold at some point, next_idx is always >= 0
-    let next_idx: number = wrap_indices.findIndex((val) => {
-      if (current_index === 0) {
-        return val < name_line_length - name_first_line_indent
-      } else {
-        return val < current_index + name_line_length
-      }
-    })
-
-    // If the next word is longer than the max line length, print it anyway
-    if (wrap_indices[next_idx] === current_index) {
-      next_idx -= 1
-    }
-
-    const next_index = wrap_indices[next_idx]
-    tspan.text(node_name.slice(current_index, next_index))
-
-    // Update variables for next line
-    if (current_index === 0) {
-      max_width = tspan.node()!.getComputedTextLength() + icon_width + node_padding
-    } else {
-      max_width = Math.max(max_width, tspan.node()!.getComputedTextLength())
-    }
-    current_index = next_index
-    title_lines += 1
-  }
-
-  class_elem.attr('y', title_lines * node_name_height)
-
-  // Find positions for potential line breaks
-  wrap_indices = [0]
-  for (const match of node_class.matchAll(line_wrap_regex)) {
-    wrap_indices.push(match.index + 1)
-  }
-  wrap_indices.push(node_class.length)
-  wrap_indices.reverse()
-
-  // Place text into multiple lines
-  current_index = 0
-  while (current_index < node_class.length) {
-    // Prepare DOM-element for next line
-    const tspan = class_elem.append('tspan')
-    tspan.attr('x', 0)
-    if (current_index > 0) {
-      tspan.attr('dy', node_class_height)
-    }
-
-    // Since this predicate is guaranteed to hold at some point, next_idx is always >= 0
-    let next_idx: number = wrap_indices.findIndex((val) => val < current_index + class_line_length)
-
-    // If the next word is longer than the max line length, print it anyway
-    if (wrap_indices[next_idx] === current_index) {
-      next_idx -= 1
-    }
-
-    const next_index = wrap_indices[next_idx]
-    tspan.text(node_class.slice(current_index, next_index))
-
-    // Update variables for next line
-    max_width = Math.max(max_width, tspan.node()!.getComputedTextLength())
-    current_index = next_index
-  }
-
-  return max_width
-}
-
-function updateNodeBody(
-  selection: d3.Selection<SVGGElement, d3.HierarchyNode<BTEditorNode>, SVGGElement, never>
-) {
-  selection.each(function (node) {
-    node.data.size.width = layoutText(this, node)
-  })
-
-  // Reset width and height of background rect
-  selection
-    .select<SVGRectElement>('.' + node_body_css_class)
-    .attr('x', null)
-    .attr('y', null)
-    .attr('width', null)
-    .attr('height', null)
-
-  // Get width and height from text content
-  selection.each(function (d) {
-    const inputs = d.data.inputs || []
-    const outputs = d.data.outputs || []
-    const max_num_grippers = Math.max(inputs.length, outputs.length)
-    const min_height =
-      io_gripper_size * max_num_grippers + io_gripper_spacing * (max_num_grippers + 1)
-    const rect = this.getBBox()
-    d.data.offset.x = rect.x - node_padding
-    d.data.offset.y = rect.y - 0.5 * node_padding
-    // Width has already been set by text layout function
-    d.data.size.width += 2 * node_padding
-    d.data.size.height = Math.max(rect.height + 1.5 * node_padding, min_height)
-  })
-
-  selection
-    .select<SVGRectElement>('.' + node_body_css_class)
-    .attr('x', (d) => d.data.offset.x)
-    .attr('y', (d) => d.data.offset.y)
-    .attr('width', (d) => d.data.size.width)
-    .attr('height', (d) => d.data.size.height)
-
-  return selection
-}
-
-watch(() => editor_store.current_tree.state, updateNodeState)
+watch(() => editor_store.tree_state_list, updateNodeState)
 function updateNodeState() {
-  if (g_vertices_ref.value === undefined) {
+  if (svg_g_ref.value === undefined) {
     console.warn('DOM is broken')
     return
   }
 
-  const g_vertex = d3.select<SVGGElement, never>(g_vertices_ref.value)
+  const vertices_group = d3.select<SVGGElement, never>(svg_g_ref.value)
 
-  const node = g_vertex
-    .selectAll<SVGSVGElement, d3.HierarchyNode<BTEditorNode>>('.' + tree_node_css_class)
+  const node = vertices_group
+    .selectAll<SVGGElement, d3.HierarchyNode<BTEditorNode>>('.' + tree_node_css_class)
     .each((node) => {
-      if (editor_store.current_tree.state === undefined) {
-        console.warn('No state to update')
-        return
-      }
-      const state = editor_store.current_tree.state.node_states.find(
-        (state) => rosToUuid(state.node_id) === node.data.node_id
+      const node_state = findNodeInTreeList(
+        editor_store.tree_state_list,
+        getNodeStates,
+        node.data.node_id
       )
-      if (state !== undefined) {
-        node.data.state = state.state
+      if (node_state !== undefined) {
+        node.data.state = node_state.state
       }
     })
   node
@@ -664,883 +310,62 @@ function updateNodeState() {
     })
 }
 
-function layoutTree(
-  selection: d3.Selection<SVGGElement, d3.HierarchyNode<BTEditorNode>, SVGGElement, never>,
-  root: d3.HierarchyNode<BTEditorNode>
-): FlextreeNode<BTEditorNode> {
-  // If the tree is in layer_mode, we have to get the max height for each layer
-  const max_height_per_layer = Array<number>(root.height + 1).fill(0.0)
-  selection.each((node: d3.HierarchyNode<BTEditorNode>) => {
-    max_height_per_layer[node.depth] = Math.max(
-      node.data.size.height,
-      max_height_per_layer[node.depth]
-    )
-  })
-
-  const tree_layout = flextree<BTEditorNode>({
-    nodeSize: (node: HierarchyNode<BTEditorNode>) => {
-      let height: number
-      if (editor_store.is_layer_mode) {
-        height = max_height_per_layer[node.depth]
-      } else {
-        height = node.data.size.height
-      }
-      height += node.depth > 0 ? node_spacing : 0
-      return [node.data.size.width, height]
-    },
-    spacing: (node: HierarchyNode<BTEditorNode>, oNode: HierarchyNode<BTEditorNode>) => {
-      if (editor_store.is_layer_mode) {
-        return node_spacing
-      }
-      if (node.parent !== oNode.parent) {
-        return 2 * node_spacing
-      } else {
-        return node_spacing
-      }
-    } // This only applies to horizontal adjacent nodes
-  })(root as HierarchyNode<BTEditorNode>)
-  //FIXME This typecast shouldn't be necessary, but apparrently the types
-  // d3.HierarchyNode and d3-hierarchy.HierarchyNode differ, as
-  // d3.HierarchyNode doesn't expose the find function???
-  // Potentially an issue with the typing library
-
-  // Bind the new data to get a selection with all flextree properties
-  selection
-    .data(
-      tree_layout.descendants().filter((node) => node.data.node_id !== uuid.NIL),
-      (node) => node.id!
-    )
-    .transition(tree_transition)
-    .attr('transform', (d: FlextreeNode<BTEditorNode>) => {
-      const x = d.x - d.data.size.width / 2.0
-      const y = d.y
-      return 'translate(' + x + ', ' + y + ')'
-    })
-
-  return tree_layout
-}
-
-function drawEdges(tree_layout: FlextreeNode<BTEditorNode>) {
-  if (g_edges_ref.value === undefined) {
-    console.warn('DOM is broken')
-    return
-  }
-
-  d3.select(g_edges_ref.value)
-    .selectAll<SVGPathElement, d3.HierarchyLink<BTEditorNode>>('.' + tree_edge_css_class)
-    .data(
-      tree_layout
-        .links()
-        .filter((link: d3.HierarchyLink<BTEditorNode>) => link.source.data.node_id !== uuid.NIL),
-      (link) => link.source.id! + '###' + link.target.id!
-    )
-    .join('path')
-    .classed(tree_edge_css_class, true) // Redundant for update elements, preserves readability
-    .transition(tree_transition)
-    .attr(
-      'd',
-      d3
-        .linkVertical<SVGPathElement, HierarchyLink<BTEditorNode>, [number, number]>()
-        .source((link: HierarchyLink<BTEditorNode>) => {
-          const source = link.source as FlextreeNode<BTEditorNode>
-          return [
-            source.x + source.data.offset.x,
-            source.y + source.data.offset.y + source.data.size.height
-          ]
-        })
-        .target((link: HierarchyLink<BTEditorNode>) => {
-          const target = link.target as FlextreeNode<BTEditorNode>
-          return [target.x + target.data.offset.x, target.y + target.data.offset.y]
-        })
-    )
-}
-
-function drawDropTargets(tree_layout: FlextreeNode<BTEditorNode>) {
-  if (g_drop_targets_ref.value === undefined) {
-    console.warn('DOM is broken')
-    return
-  }
-
-  const drop_targets: DropTarget[] = []
-
-  // Only draw drop targets if not displaying a subtree
-  if (!editor_store.has_selected_subtree) {
-    // Construct the list of drop targets that should exist
-    tree_layout.each((node: FlextreeNode<BTEditorNode>) => {
-      if (node.data.node_id === uuid.NIL) {
-        return
-      }
-
-      // Draw all left and right targets, even though they sometimes overlap
-      // because it looks odd if they're missing once nodes are spaced out further
-      drop_targets.push({ node: node, position: Position.LEFT })
-      drop_targets.push({ node: node, position: Position.CENTER })
-      drop_targets.push({ node: node, position: Position.RIGHT })
-      drop_targets.push({ node: node, position: Position.TOP }) // Does this make sense?
-      if (node.data.max_children === -1 || node.data.max_children > node.data.child_ids.length) {
-        drop_targets.push({ node: node, position: Position.BOTTOM })
-      }
-    })
-  }
-
-  // If there are no drop targets to draw, draw one at root
-  // Adjust the size of the drop target here. Don't do this on init, it ruins the tree layout
-  if (drop_targets.length === 0) {
-    drop_targets.push({ node: tree_layout, position: Position.ROOT })
-  }
-
-  // Join those with the existing drop targets and draw them
-  d3.select(g_drop_targets_ref.value)
-    .selectAll<SVGRectElement, DropTarget>('.' + drop_target_css_class)
-    .data(drop_targets, (d) => d.node.id! + '###' + d.position)
-    .join('rect')
-    .classed(drop_target_css_class, true)
-    .attr('width', (d) => {
-      // Switch cases without breaks work like a convenient OR
-      switch (d.position) {
-        case Position.LEFT:
-        case Position.RIGHT:
-          return node_spacing
-        case Position.TOP:
-        case Position.BOTTOM:
-        case Position.CENTER:
-          return d.node.data.size.width
-        case Position.ROOT:
-          return drop_target_root_size
-        default:
-          return 0
-      }
-    })
-    .attr('height', (d) => {
-      switch (d.position) {
-        case Position.LEFT:
-        case Position.RIGHT:
-        case Position.CENTER:
-          return d.node.data.size.height
-        case Position.TOP:
-        case Position.BOTTOM:
-          return 0.5 * node_spacing
-        case Position.ROOT:
-          return drop_target_root_size
-        default:
-          return 0
-      }
-    })
-    .attr('x', (d) => {
-      if (d.position === Position.ROOT) {
-        return -drop_target_root_size / 2
-      }
-      let x = d.node.x + d.node.data.offset.x
-      if (d.position === Position.RIGHT) {
-        x += 0.5 * d.node.data.size.width
-      } else {
-        x -= 0.5 * d.node.data.size.width
-        if (d.position === Position.LEFT) {
-          x -= node_spacing
-        }
-      }
-      return x
-    })
-    .attr('y', (d) => {
-      if (d.position === Position.ROOT) {
-        return -0.5 * node_spacing
-      }
-      let y = d.node.y + d.node.data.offset.y
-      if (d.position === Position.TOP) {
-        y -= 0.5 * node_spacing
-      }
-      if (d.position === Position.BOTTOM) {
-        y += d.node.data.size.height
-      }
-      return y
-    })
-    .attr('opacity', 0.2)
-    .on('mouseover.highlight', function () {
-      d3.select(this).attr('opacity', 0.8)
-    })
-    .on('mouseout.highlight', function () {
-      d3.select(this).attr('opacity', 0.2)
-    })
-    .on('mouseup.dragdrop', (ev, d) => {
-      if (editor_store.dragging_new_node) {
-        addNewNode(d)
-      } else if (editor_store.dragging_existing_node) {
-        moveExistingNode(d)
-      } else {
-        console.warn('Unintended drag release')
-      }
-    })
-}
-
-function dropTargetGroupVisibility(): string {
-  if (
-    editor_store.dragging_new_node !== undefined ||
-    editor_store.dragging_existing_node !== undefined
-  ) {
-    return 'visible'
-  }
-  return 'hidden'
-}
-
-async function moveExistingNode(drop_target: DropTarget) {
-  if (editor_store.dragging_existing_node === undefined) {
-    console.warn('Tried to move existing node by dragging but none selected')
-    return
-  }
-
-  const target_node_id = editor_store.dragging_existing_node.data.node_id
-  const target_node_name = editor_store.dragging_existing_node.data.name
-
-  if (!drop_target.node.parent) {
-    console.error('A tree with an existing node should never show the root target')
-    return
-  }
-
-  if (drop_target.position === Position.BOTTOM) {
-    await moveNode(target_node_id, target_node_name, drop_target.node.data.node_id, 0)
-    return
-  }
-
-  if (drop_target.position === Position.LEFT || drop_target.position === Position.RIGHT) {
-    const parent_node_id = drop_target.node.parent.data.node_id
-    let index = drop_target.node.parent.children!.indexOf(drop_target.node)
-    if (drop_target.position === Position.RIGHT) {
-      index++
-    }
-    // If the node is moved in it's own row (same parent), we need to offset the index
-    if (
-      parent_node_id === editor_store.dragging_existing_node.parent!.data.node_id &&
-      index >
-        drop_target.node.parent.children!.findIndex(
-          (node: FlextreeNode<BTEditorNode>) => node.data.node_id === target_node_id
-        )
-    ) {
-      index--
-    }
-    await moveNode(target_node_id, target_node_name, parent_node_id, index)
-    return
-  }
-
-  // Checks on whether the new node is an appropriate replacement/ancestor
-  // are presumed done based on whether this target was available in the first place.
-
-  if (drop_target.position === Position.TOP) {
-    const parent_node_id = drop_target.node.parent.data.node_id
-    let index = drop_target.node.parent.children!.indexOf(drop_target.node)
-    // If the node is moved in it's own row (same parent), we need to offset the index
-    if (
-      parent_node_id === editor_store.dragging_existing_node.parent!.data.node_id &&
-      index >
-        drop_target.node.parent.children!.findIndex(
-          (node: FlextreeNode<BTEditorNode>) => node.data.node_id === target_node_id
-        )
-    ) {
-      index--
-    }
-
-    // Care has to be taken regarding order of operations to not overload the parent node.
-    // Insert new node at the end, then move node to old position
-    await moveNode(drop_target.node.data.node_id, drop_target.node.data.name, target_node_id, -1)
-    await moveNode(target_node_id, target_node_name, parent_node_id, index)
-    return
-  }
-
-  if (drop_target.position === Position.CENTER) {
-    await replaceNode(
-      drop_target.node.data.node_id,
-      drop_target.node.data.name,
-      target_node_id,
-      target_node_name
-    )
-    return
-  }
-
-  console.warn("Couldn't identify drop target position, this should never happen")
-}
-
-watchEffect(toggleExistingNodeTargets)
-function toggleExistingNodeTargets() {
-  if (g_drop_targets_ref.value === undefined) {
+watch(
+  [() => editor_store.dragging_existing_node, () => editor_store.dragging_new_node],
+  toggleValidDropTargets
+)
+function toggleValidDropTargets() {
+  if (svg_g_ref.value === undefined) {
     console.warn('DOM is broken')
     return
   }
 
   //Reset visibility for all targets
-  const targets = d3
-    .select(g_drop_targets_ref.value)
+  d3.select(svg_g_ref.value)
     .selectAll<SVGRectElement, DropTarget>('.' + drop_target_css_class)
-    .attr('visibility', null)
-
-  if (editor_store.dragging_existing_node === undefined) {
-    return
-  }
-
-  // The first filter hides all nodes in the currently dragged subtree
-  // The second filter hides all nodes where dropping would overload a child node count
-  //TODO Maybe we should hide targets that would place the node in the same spot
-
-  // If the filter returns true (keeps the node) it gets hidden
-  targets
-    .filter((drop_target: DropTarget) => {
-      return (
-        editor_store
-          .dragging_existing_node!.descendants()
-          .find((node: d3.HierarchyNode<BTEditorNode>) => {
-            return node.data.node_id === drop_target.node.data.node_id
-          }) !== undefined
-      )
-    })
     .attr('visibility', 'hidden')
 
-  // If the filter returns true (keeps the node) it gets hidden
-  targets
-    .filter((drop_target: DropTarget) => {
-      switch (drop_target.position) {
-        case Position.CENTER:
-          return (
-            editor_store.dragging_existing_node!.data.max_children !== -1 &&
-            drop_target.node.data.child_ids.length +
-              editor_store.dragging_existing_node!.data.child_ids.length >
-              editor_store.dragging_existing_node!.data.max_children
-          )
-        case Position.TOP:
-          return (
-            editor_store.dragging_existing_node!.data.max_children !== -1 &&
-            editor_store.dragging_existing_node!.data.child_ids.length >=
-              editor_store.dragging_existing_node!.data.max_children
-          )
-        case Position.BOTTOM:
-          return (
-            drop_target.node.data.max_children !== -1 &&
-            drop_target.node.data.node_id !==
-              editor_store.dragging_existing_node!.parent!.data.node_id &&
-            drop_target.node.data.child_ids.length >= drop_target.node.data.max_children
-          )
-        case Position.LEFT:
-        case Position.RIGHT:
-          return (
-            drop_target.node.parent!.data.max_children !== -1 &&
-            drop_target.node.parent!.data.node_id !==
-              editor_store.dragging_existing_node!.parent!.data.node_id &&
-            drop_target.node.parent!.data.child_ids.length >=
-              drop_target.node.parent!.data.max_children
-          )
-        case Position.ROOT:
-        // This should never happen, as an existing node implies a non-empty tree
-        default:
-          return true
-      }
-    })
-    .attr('visibility', 'hidden')
+  if (tree_display === undefined) {
+    console.warn('Tree is not drawn')
+    return
+  }
+
+  if (editor_store.dragging_existing_node !== undefined) {
+    tree_display.toggleExistingNodeDropTargets(editor_store.dragging_existing_node)
+  }
+  if (editor_store.dragging_new_node !== undefined) {
+    tree_display.toggleNewNodeDropTargets(editor_store.dragging_new_node)
+  }
 }
 
-async function addNewNode(drop_target: DropTarget) {
-  if (editor_store.dragging_new_node === undefined) {
-    console.warn('Tried to add new node by dragging but none selected')
-    return
-  }
-
-  const msg = buildNodeMessage(editor_store.dragging_new_node)
-
-  if (drop_target.position === Position.ROOT) {
-    await addNode(msg, uuid.NIL, 0)
-    return
-  }
-
-  if (!drop_target.node.parent) {
-    console.error('All non-root targets should have a set parent node')
-    return
-  }
-
-  if (drop_target.position === Position.BOTTOM) {
-    const parent_node_id = drop_target.node.data.node_id
-    await addNode(msg, parent_node_id, 0)
-    return
-  }
-
-  if (drop_target.position === Position.LEFT || drop_target.position === Position.RIGHT) {
-    const parent_node_id = drop_target.node.parent.data.node_id
-    let index = drop_target.node.parent.children!.indexOf(drop_target.node)
-    if (drop_target.position === Position.RIGHT) {
-      index++
-    }
-    await addNode(msg, parent_node_id, index)
-    return
-  }
-
-  // Checks on whether the new node is an appropriate replacement/ancestor
-  // are presumed done based on whether this target was available in the first place.
-
-  if (drop_target.position === Position.TOP) {
-    const parent_node_id = drop_target.node.parent.data.node_id
-    const index = drop_target.node.parent.children!.indexOf(drop_target.node)
-
-    // Care has to be taken regarding order of operations to not overload the parent node.
-    // Insert at top temporarily
-    const new_node_id = await addNode(msg, uuid.NIL, -1)
-    await moveNode(drop_target.node.data.node_id, drop_target.node.data.name, new_node_id, 0)
-    await moveNode(new_node_id, msg.name, parent_node_id, index)
-    return
-  }
-
-  if (drop_target.position === Position.CENTER) {
-    const new_node_id = await addNode(msg, uuid.NIL, -1)
-    await replaceNode(
-      drop_target.node.data.node_id,
-      drop_target.node.data.name,
-      new_node_id,
-      msg.name
-    )
-    return
-  }
-
-  console.warn("Couldn't identify drop target position, this should never happen")
-}
-
-watchEffect(toggleNewNodeTargets)
-function toggleNewNodeTargets() {
-  if (g_drop_targets_ref.value === undefined) {
-    console.warn('DOM is broken')
-    return
-  }
-
-  //Reset visibility for all targets
-  const targets = d3
-    .select(g_drop_targets_ref.value)
-    .selectAll<SVGRectElement, DropTarget>('.' + drop_target_css_class)
-    .attr('visibility', null)
-
-  if (editor_store.dragging_new_node === undefined) {
-    return
-  }
-
-  // If the filter returns true (keeps the node) it gets hidden
-  targets
-    .filter((drop_target: DropTarget) => {
-      switch (drop_target.position) {
-        case Position.CENTER:
-          return (
-            editor_store.dragging_new_node!.max_children !== -1 &&
-            drop_target.node.data.child_ids.length > editor_store.dragging_new_node!.max_children
-          )
-        case Position.TOP:
-          return editor_store.dragging_new_node!.max_children === 0
-        case Position.BOTTOM:
-          return (
-            drop_target.node.data.max_children !== -1 &&
-            drop_target.node.data.child_ids.length >= drop_target.node.data.max_children
-          )
-        case Position.LEFT:
-        case Position.RIGHT:
-          return (
-            drop_target.node.parent!.data.max_children !== -1 &&
-            drop_target.node.parent!.data.child_ids.length >=
-              drop_target.node.parent!.data.max_children
-          )
-        case Position.ROOT:
-          return false
-        default:
-          return true
-      }
-    })
-    .attr('visibility', 'hidden')
-}
-
-function drawDataGraph(tree_layout: FlextreeNode<BTEditorNode>, data_wirings: Wiring[]) {
-  if (g_data_graph_ref.value === undefined || g_data_vertices_ref.value === undefined) {
-    console.warn('DOM is broken')
-    return
-  }
-
-  const data_points: DataEdgeTerminal[] = []
-
-  tree_layout.each((node: FlextreeNode<BTEditorNode>) => {
-    if (node.data.node_id === uuid.NIL) {
-      return
-    }
-
-    node.data.inputs.map((input: TrimmedNodeData, index: number) => {
-      data_points.push({
-        node: node,
-        index: index,
-        kind: IOKind.INPUT,
-        key: input.key,
-        type: input.serialized_type,
-        x: node.x + node.data.offset.x - node.data.size.width * 0.5 - io_gripper_size,
-        y:
-          node.y +
-          node.data.offset.y +
-          io_gripper_spacing +
-          index * (io_gripper_size + io_gripper_spacing)
-      })
-    })
-    node.data.outputs.map((output: TrimmedNodeData, index: number) => {
-      data_points.push({
-        node: node,
-        index: index,
-        kind: IOKind.OUTPUT,
-        key: output.key,
-        type: output.serialized_type,
-        x: node.x + node.data.offset.x + node.data.size.width * 0.5,
-        y:
-          node.y +
-          node.data.offset.y +
-          io_gripper_spacing +
-          index * (io_gripper_size + io_gripper_spacing)
-      })
-    })
-  })
-
-  const g_data_vertices = d3
-    .select(g_data_vertices_ref.value)
-    .selectAll<SVGGElement, DataEdgeTerminal>('.' + data_vert_group_css_class)
-    .data(data_points, (d) => d.node.id! + '###' + d.kind + '###' + d.key)
-    .join(drawNewDataVert)
-
-  if (!editor_store.has_selected_subtree) {
-    //NOTE These are added here, because they see an outdated datum otherwise
-    // d3 claims that listeners are passed a "current" datum,
-    // but that appears to be wrong.
-    g_data_vertices
-      .select('.' + data_vert_grip_css_class)
-      .on('mousedown.drawedge', (ev, term: DataEdgeTerminal) => {
-        editor_store.startDrawingDataEdge(term)
-      })
-      .on('mouseup.drawedge', (ev, term: DataEdgeTerminal) => {
-        if (editor_store.data_edge_endpoint === undefined) {
-          console.warn('Unintended data edge draw')
-          return
-        }
-
-        if (!typesCompatible(term, editor_store.data_edge_endpoint)) {
-          console.warn('Invalid edge')
-          return
-        }
-
-        if (term.kind === IOKind.INPUT) {
-          addNewDataEdge(editor_store.data_edge_endpoint, term)
-        } else {
-          addNewDataEdge(term, editor_store.data_edge_endpoint)
-        }
-      })
-  }
-
-  // Since types of DataVerts can change, type values are added out here
-  g_data_vertices
-    .select('.' + data_vert_label_css_class)
-    .select('.' + data_vert_label_type_css_class)
-    .text((d) => '(type: ' + prettyprint_type(d.type) + ')')
-
-  g_data_vertices
-    .transition(tree_transition)
-    //NOTE group elements can't be positioned with x= and y=
-    .attr('transform', (d) => 'translate(' + d.x + ', ' + d.y + ')')
-
-  drawDataEdges(data_points, g_data_vertices, data_wirings)
-}
-
-function drawNewDataVert(
-  selection: d3.Selection<d3.EnterElement, DataEdgeTerminal, SVGGElement, unknown>
-) {
-  const groups = selection
-    .append('g')
-    .classed(data_vert_group_css_class, true)
-    .on('mouseover.highlight', function () {
-      if (editor_store.is_dragging) {
-        // Highlight compatible vertices when dragging
-        const compat = d3.select(this).classed(data_graph_comaptible_css_class)
-        d3.select(this).classed(data_graph_hover_css_class, compat)
-        d3.select(draw_indicator_ref.value!).classed(data_graph_hover_css_class, compat)
-      } else {
-        d3.select(this)
-          .classed(data_graph_hover_css_class, true)
-          .attr('id', data_vert1_highlight_css_id)
-          .select('.' + data_vert_label_css_class)
-          .attr('visibility', 'visible')
-      }
-    })
-    .on('mouseout.highlight', function () {
-      if (editor_store.is_dragging) {
-        d3.select(this).classed(data_graph_hover_css_class, false)
-        d3.select(draw_indicator_ref.value!).classed(data_graph_hover_css_class, false)
-      } else {
-        d3.select(this)
-          .classed(data_graph_hover_css_class, false)
-          .attr('id', null)
-          .select('.' + data_vert_label_css_class)
-          .attr('visibility', 'hidden')
-      }
-    })
-
-  groups
-    .append('rect')
-    .classed(data_vert_grip_css_class, true)
-    .attr('width', io_gripper_size)
-    .attr('height', io_gripper_size)
-
-  const labels = groups
-    .append('text')
-    .classed(data_vert_label_css_class, true)
-    .attr('dominant-baseline', 'middle')
-    .attr('visibility', 'hidden')
-    .attr('text-anchor', (d) => (d.kind === IOKind.INPUT ? 'end' : 'start'))
-    .text((d) => d.key)
-    .attr('x', (d) => {
-      switch (d.kind) {
-        case IOKind.INPUT:
-          return -5
-        case IOKind.OUTPUT:
-          return io_gripper_size + 5
-        case IOKind.OTHER:
-        default:
-          return 0
-      }
-    })
-    .attr('y', 0.5 * io_gripper_size)
-
-  labels
-    .append('tspan')
-    .classed(data_vert_label_type_css_class, true)
-    .attr('x', function () {
-      return d3.select(this.parentElement).attr('x')
-    })
-    .attr('dy', '1em') // Space out 2nd line
-
-  // The join pattern requires a return of the appended elements
-  return groups
-}
-
-function drawDataEdges(
-  data_points: DataEdgeTerminal[],
-  g_data_vertices: d3.Selection<SVGGElement, DataEdgeTerminal, SVGGElement, unknown>,
-  data_wirings: Wiring[]
-) {
-  if (
-    g_data_graph_ref.value === undefined ||
-    g_data_vertices_ref.value === undefined ||
-    g_data_edges_ref.value === undefined
-  ) {
-    console.warn('DOM is broken')
-    return
-  }
-
-  // Construct edge array by matching tree_msg wirings
-  const data_edges: DataEdge[] = []
-
-  function matchEndpoint(wire_point: NodeDataLocation, terminal: DataEdgeTerminal): boolean {
-    return (
-      rosToUuid(wire_point.node_id) === terminal.node.data.node_id &&
-      wire_point.data_kind === terminal.kind &&
-      wire_point.data_key === terminal.key
-    )
-  }
-
-  data_wirings.forEach((wiring: Wiring) => {
-    // Match Terminals with wiring data
-    const source = data_points.find((term: DataEdgeTerminal) => matchEndpoint(wiring.source, term))
-    const target = data_points.find((term: DataEdgeTerminal) => matchEndpoint(wiring.target, term))
-
-    if (source === undefined || target === undefined) {
-      console.warn('Bad data edge', source, target)
-      return
-    }
-
-    // Try to assign output as source and input as target
-    if (source.kind === IOKind.INPUT) {
-      data_edges.push({
-        source: target,
-        target: source,
-        wiring: wiring
-      })
-    } else {
-      data_edges.push({
-        source: source,
-        target: target,
-        wiring: wiring
-      })
-    }
-  })
-
-  d3.select(g_data_edges_ref.value)
-    .selectAll<SVGPathElement, DataEdge>('.' + data_edge_css_class)
-    .data(
-      data_edges,
-      (d: DataEdge) =>
-        d.source.node.id! +
-        '###' +
-        d.source.kind +
-        '###' +
-        d.source.index +
-        '#####' +
-        d.target.node.id! +
-        '###' +
-        d.target.kind +
-        '###' +
-        d.target.index
-    )
-    .join('path')
-    .classed(data_edge_css_class, true)
-    .on('click.select', (event, edge: DataEdge) => {
-      editor_store.selectEdge(edge.wiring)
-      event.stopPropagation()
-    })
-    .on('mouseover.highlight', function (ev, edge: DataEdge) {
-      if (editor_store.is_dragging) {
-        return // No highlights while dragging
-      }
-      g_data_vertices
-        .filter((term: DataEdgeTerminal) => term === edge.source || term === edge.target)
-        .dispatch('mouseover')
-        .attr('id', (term: DataEdgeTerminal) =>
-          term.kind === IOKind.INPUT ? data_vert2_highlight_css_id : data_vert1_highlight_css_id
-        )
-        //Hide target label
-        .select('.' + data_vert_label_css_class)
-        .attr('visibility', (term: DataEdgeTerminal) => {
-          if (term.kind === IOKind.INPUT) {
-            return 'hidden'
-          }
-          return 'visible'
-        })
-
-      d3.select(this)
-        .classed(data_graph_hover_css_class, true)
-        .attr('id', data_edge_highlight_css_id)
-    })
-    .on('mouseout.highlight', function (ev, edge: DataEdge) {
-      if (editor_store.is_dragging) {
-        return // No highlights while dragging
-      }
-      g_data_vertices
-        .filter((term: DataEdgeTerminal) => term === edge.source || term === edge.target)
-        .dispatch('mouseout')
-      d3.select(this).classed(data_graph_hover_css_class, false).attr('id', null)
-    })
-    .transition(tree_transition)
-    .attr('d', (edge: DataEdge) => drawDataLine(edge.source, edge.target))
-}
-
-function drawDataLine(source: DataEdgePoint, target: DataEdgePoint) {
-  const lineGen = d3
-    .line<DataEdgePoint>()
-    .x((p) => p.x + io_gripper_size / 2)
-    .y((p) => p.y + io_gripper_size / 2)
-    .curve(d3.curveCatmullRom.alpha(0.9))
-  let y_offset = 0
-  if (Math.abs(source.y - target.y) < io_edge_bump_thresh) {
-    y_offset = Math.min(source.y, target.y) - io_edge_bump_factor * Math.abs(source.x - target.x)
-  }
-  const source_offset: DataEdgePoint = {
-    x: source.x + io_edge_offset,
-    y: source.y
-  }
-  const target_offset: DataEdgePoint = {
-    x: target.x - io_edge_offset,
-    y: target.y
-  }
-  const midpoint: DataEdgePoint = {
-    x: (source.x + target.x) / 2,
-    y: y_offset ? y_offset : (source.y + target.y) / 2
-  }
-  // Backwards edges require some extra work
-  if (source.x > target.x) {
-    if (y_offset === 0) {
-      source_offset.y +=
-        io_edge_curve_factor * (target.y - source.y) +
-        Math.sign(target.y - source.y) * io_edge_curve_offset
-      target_offset.y +=
-        io_edge_curve_factor * (source.y - target.y) +
-        Math.sign(source.y - target.y) * io_edge_curve_offset
-    } else {
-      const curve_offset =
-        io_edge_curve_offset + io_edge_curve_factor * Math.abs(source.x - target.x)
-      source_offset.y -= curve_offset
-      midpoint.y -= curve_offset * 4
-      target_offset.y -= curve_offset
-    }
-  }
-  return lineGen([source, source_offset, midpoint, target_offset, target])
-}
-
-watchEffect(toggleDataEdgeTargets)
+watch(() => editor_store.data_edge_endpoint, toggleDataEdgeTargets)
 function toggleDataEdgeTargets() {
-  if (g_data_vertices_ref.value === undefined || draw_indicator_ref.value === undefined) {
+  if (svg_g_ref.value === undefined || draw_indicator_ref.value === undefined) {
     console.warn('DOM is broken')
     return
   }
 
   // Reset visibility on all grippers
   const data_verts = d3
-    .select(g_data_vertices_ref.value)
+    .select(svg_g_ref.value)
     .selectAll<SVGGElement, DataEdgeTerminal>('.' + data_vert_group_css_class)
     .classed(data_graph_comaptible_css_class, false)
 
   data_verts.select('.' + data_vert_label_css_class).attr('visibility', 'hidden')
 
   // Reset drawing indicator
-  const draw_path = d3.select(draw_indicator_ref.value).attr('d', null)
+  d3.select(draw_indicator_ref.value).attr('d', null)
 
   if (editor_store.data_edge_endpoint === undefined) {
     return
   }
 
-  data_verts
-    .filter((term: DataEdgeTerminal) => typesCompatible(term, editor_store.data_edge_endpoint!))
-    .classed(data_graph_comaptible_css_class, true)
+  if (tree_display === undefined) {
+    console.warn('Tree is not drawn')
+    return
+  }
 
-  draw_path.attr('d', () =>
-    drawDataLine(editor_store.data_edge_endpoint!, editor_store.data_edge_endpoint!)
-  )
-}
-
-// Caller needs to ensure to pass the input as source
-function addNewDataEdge(source: DataEdgeTerminal, target: DataEdgeTerminal) {
-  ros_store.wire_data_service.callService(
-    {
-      wirings: [
-        {
-          source: {
-            node_id: uuidToRos(source.node.data.node_id),
-            data_kind: source.kind,
-            data_key: source.key
-          },
-          target: {
-            node_id: uuidToRos(target.node.data.node_id),
-            data_kind: target.kind,
-            data_key: target.key
-          }
-        } as Wiring
-      ],
-      ignore_failure: false //TODO what does this do?
-    } as WireNodeDataRequest,
-    (response: WireNodeDataResponse) => {
-      const source_name = editor_store.current_tree.structure!.nodes.find(
-        (node) => rosToUuid(node.node_id) === source.node.data.node_id
-      )!.name
-      const target_name = editor_store.current_tree.structure!.nodes.find(
-        (node) => rosToUuid(node.node_id) === target.node.data.node_id
-      )!.name
-      if (response.success) {
-        notify({
-          title: 'Added data edge: ' + source_name + ' -> ' + target_name + '!',
-          type: 'success'
-        })
-      } else {
-        notify({
-          title: 'Failed to add data edge: ' + source_name + ' -> ' + target_name + '!',
-          text: response.error_message,
-          type: 'warn'
-        })
-      }
-    },
-    (error: string) => {
-      notify({
-        title: 'Failed to call dataWiring service',
-        text: error,
-        type: 'error'
-      })
-    }
-  )
+  tree_display.highlightCompatibleVertices(editor_store.data_edge_endpoint)
 }
 
 watch(
@@ -1548,7 +373,7 @@ watch(
   () => colorSelectedNodes()
 )
 function colorSelectedNodes() {
-  if (g_vertices_ref.value === undefined) {
+  if (svg_g_ref.value === undefined) {
     console.warn('DOM is broken')
     return
   }
@@ -1562,7 +387,7 @@ function colorSelectedNodes() {
   )
   const all_selected_nodes = old_selected_nodes.concat(new_selected_nodes)
 
-  d3.select<SVGGElement, never>(g_vertices_ref.value)
+  d3.select<SVGGElement, never>(svg_g_ref.value)
     .selectAll<SVGSVGElement, FlextreeNode<BTEditorNode>>('.' + tree_node_css_class)
     .select<SVGRectElement>('.' + node_body_css_class)
     .classed(node_selected_css_class, (node: FlextreeNode<BTEditorNode>) =>
@@ -1577,24 +402,23 @@ watch(
   }
 )
 function colorSelectedEdge() {
-  if (g_data_edges_ref.value === undefined || g_data_vertices_ref.value === undefined) {
+  if (svg_g_ref.value === undefined) {
     console.warn('DOM is broken')
     return
   }
 
-  const g_data_edges = d3.select(g_data_edges_ref.value)
-  const g_data_vertices = d3.select(g_data_vertices_ref.value)
+  const root_element = d3.select(svg_g_ref.value)
 
-  g_data_edges
+  root_element
     .selectAll<SVGPathElement, DataEdge>('.' + data_edge_css_class)
     .classed(data_graph_select_css_class, false)
 
-  g_data_vertices
+  root_element
     .selectAll<SVGGElement, DataEdgeTerminal>('.' + data_vert_group_css_class)
     .classed(data_graph_select_css_class, false)
 
   if (editor_store.selected_edge !== undefined) {
-    const newEdge: DataEdge = g_data_edges
+    const newEdge: DataEdge = root_element
       .selectAll<SVGPathElement, DataEdge>('.' + data_edge_css_class)
       .filter((edge: DataEdge) => {
         return edge.wiring === editor_store.selected_edge
@@ -1602,7 +426,7 @@ function colorSelectedEdge() {
       .classed(data_graph_select_css_class, true)
       .datum()
 
-    g_data_vertices
+    root_element
       .selectAll<SVGGElement, DataEdgeTerminal>('.' + data_vert_group_css_class)
       .filter((term: DataEdgeTerminal) => {
         return term === newEdge.source || term === newEdge.target
@@ -1672,11 +496,7 @@ onMounted(() => {
 
   // show the selection rectangle on mousemove
   viewport.on('mousemove.drawselect', (event) => {
-    if (
-      svg_g_ref.value === undefined ||
-      g_vertices_ref.value === undefined ||
-      selection_rect_ref.value === undefined
-    ) {
+    if (svg_g_ref.value === undefined || selection_rect_ref.value === undefined) {
       console.warn('DOM is broken')
       return
     }
@@ -1701,9 +521,9 @@ onMounted(() => {
       const temp_selected_nodes: string[] = []
 
       // Update which nodes are in the selection
-      d3.select<SVGGElement, never>(g_vertices_ref.value)
-        .selectAll<SVGForeignObjectElement, FlextreeNode<BTEditorNode>>('.' + tree_node_css_class)
-        .select<HTMLBodyElement>('.' + node_body_css_class)
+      d3.select<SVGGElement, never>(svg_g_ref.value)
+        .selectAll<SVGGElement, FlextreeNode<BTEditorNode>>('.' + tree_node_css_class)
+        .select<SVGRectElement>('.' + node_body_css_class)
         .each((node: FlextreeNode<BTEditorNode>) => {
           // Select all nodes in the selection rectangle
           // Node coordinates are given for the top-center point
@@ -1723,7 +543,7 @@ onMounted(() => {
 
   // detect the selected nodes on mouseup
   viewport.on('mouseup.drawselect', () => {
-    if (g_vertices_ref.value === undefined || selection_rect_ref.value === undefined) {
+    if (svg_g_ref.value === undefined || selection_rect_ref.value === undefined) {
       console.warn('DOM is broken')
       return
     }
@@ -1782,25 +602,17 @@ onMounted(() => {
 <template>
   <svg id="editor_viewport" ref="viewport_ref" class="reactive-svg" :class="editor_store.skin">
     <g id="container" ref="svg_g_ref">
-      <g ref="g_edges_ref" class="edges" />
-      <g ref="g_vertices_ref" class="vertices" />
-      <g
-        ref="g_data_graph_ref"
-        class="data_graph"
-        :visibility="editor_store.show_data_graph ? 'visible' : 'hidden'"
-      >
-        <g ref="g_data_edges_ref" class="data_edges" />
-        <g ref="g_data_vertices_ref" class="data_vertices" />
-        <!--Below is used to pull elements to the foreground on hover-->
-        <use :href="'#' + data_vert1_highlight_css_id" pointer-events="none" />
-        <use :href="'#' + data_vert2_highlight_css_id" pointer-events="none" />
-        <use :href="'#' + data_edge_highlight_css_id" pointer-events="none" />
-      </g>
-      <g ref="g_drop_targets_ref" class="drop_targets" :visibility="dropTargetGroupVisibility()" />
+      <g ref="tree_root_ref" />
+
+      <!--Below is used to pull elements to the foreground on hover-->
+      <use :href="'#' + data_vert1_highlight_css_id" pointer-events="none" />
+      <use :href="'#' + data_vert2_highlight_css_id" pointer-events="none" />
+      <use :href="'#' + data_edge_highlight_css_id" pointer-events="none" />
 
       <path ref="draw_indicator_ref" class="drawing-indicator" />
       <rect ref="selection_rect_ref" class="selection" />
     </g>
+
     <text
       x="10"
       y="20"
